sub:int and refactor
