+++ src/audit/models.py
@@ -1,5 +1,6 @@
 from datetime import datetime
 from gino.ext.starlette import Gino
+from pydantic import BaseModel
 import sqlalchemy
 from sqlalchemy import Column, DateTime, Integer, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import ARRAY, TIMESTAMP
@@ -19,18 +20,57 @@ db = Gino(
 )
 
 
+# base audit log class
+
+
 class AuditLog(db.Model):
     request_url = Column(String, nullable=False)
     status_code = Column(Integer, nullable=False)
     timestamp = Column(DateTime, nullable=False, default=sqlalchemy.func.now())
     username = Column(String, nullable=False)
-    sub = Column(String, nullable=False)
+    sub = Column(Integer, nullable=True)  # can be null for public data
+
+
+class CreateLogInput(BaseModel):
+    request_url: str
+    status_code: int
+    username: str
+    # timestamp: we store DateTimes in the DB but the API takes
+    # int timestamps as input
+    timestamp: int = None
+    sub: int
+
+
+# child audit log classes
 
 
 class PresignedUrl(AuditLog):
     __tablename__ = "presigned_url"
 
-    guid = Column(String, primary_key=True)
+    guid = Column(String, nullable=False)
     resource_paths = Column(ARRAY(String), nullable=False)
     action = Column(String, nullable=False)
-    protocol = Column(String, nullable=True)
+    protocol = Column(String, nullable=True)  # can be null if action=="upload"
+
+
+class CreatePresignedUrlLogInput(CreateLogInput):
+    guid: str
+    resource_paths: list
+    action: str
+    protocol: str = None
+
+
+class Login(AuditLog):
+    __tablename__ = "login"
+
+    idp = Column(String, nullable=False)
+    fence_idp = Column(String, nullable=True)
+    shib_idp = Column(String, nullable=True)
+    client_id = Column(String, nullable=False)
+
+
+class CreateLoginLogInput(CreateLogInput):
+    idp: str
+    fence_idp: str = None
+    shib_idp: str = None
+    client_id: str
